+++ test/meal.e2e-spec.ts
@@ -9,15 +9,17 @@ import { JwtManagerService } from '../src/modules/jwt-manager/jwt-manager.servic
 import { RedisService } from '../src/modules/redis/redis.service';
 import { MealRepository } from '../src/mongodb/repositories/meal.repository';
 import { SpoonacularApiService } from '../src/modules/api/spoonacular/spoonacular.api.service';
-import { DetailedMeal, DetailedMealWithTranslations, RatedMeal } from '../src/modules/meal/meal.types';
+import { DetailedMeal, DetailedMealWithTranslations, MealRating, RatedMeal } from '../src/modules/meal/meal.types';
 import { SearchQueryRepository } from '../src/mongodb/repositories/search-query.repository';
 import { MealCommentRepository } from '../src/mongodb/repositories/meal-comment.repository';
+import { MealRatingRepository } from '../src/mongodb/repositories/meal-rating.repository';
 
 describe('UserController (e2e)', () => {
     let app: INestApplication;
     let mealService: MealService;
     let mealRepository: MealRepository;
     let mealCommentRepository: MealCommentRepository;
+    let mealRatingRepository: MealRatingRepository;
     let searchQueryRepository: SearchQueryRepository;
     let jwtManagerService: JwtManagerService;
     let redisService: RedisService;
@@ -48,6 +50,10 @@ describe('UserController (e2e)', () => {
         findOne: () => {},
         findById: () => {}
     };
+    const mockMealRatingRepositoryProvider = {
+        ...mockMealRepositoryProvider,
+        getAverageRatingForMeal: jest.fn()
+    };
     const redisServiceProvider = {
         set: jest.fn(),
         get: jest.fn(),
@@ -73,6 +79,7 @@ describe('UserController (e2e)', () => {
             .overrideProvider(LoggerService).useValue(loggerServiceProvider)
             .overrideProvider(MealRepository).useValue(mockMealRepositoryProvider)
             .overrideProvider(MealCommentRepository).useValue(mockMealRepositoryProvider)
+            .overrideProvider(MealRatingRepository).useValue(mockMealRatingRepositoryProvider)
             .overrideProvider(SearchQueryRepository).useValue(mockMealRepositoryProvider)
             .overrideProvider(RedisService).useValue(redisServiceProvider)
             .overrideProvider(JwtManagerService).useValue(jwtManagerServiceProvider)
@@ -86,6 +93,7 @@ describe('UserController (e2e)', () => {
         mealService = moduleRef.get(MealService);
         mealRepository = moduleRef.get(MealRepository);
         mealCommentRepository = moduleRef.get(MealCommentRepository);
+        mealRatingRepository = moduleRef.get(MealRatingRepository);
         searchQueryRepository = moduleRef.get(SearchQueryRepository);
         jwtManagerService = moduleRef.get(JwtManagerService);
         redisService = moduleRef.get(RedisService);
@@ -220,6 +228,26 @@ describe('UserController (e2e)', () => {
         });
     });
 
+    describe('/meals/:id/rating (GET)', () => {
+        it('should receive all rating for a particular meal', () => {
+            const mockMealId = 'mock meal id';
+            const mockHasMeal: boolean = true;
+            const mockMealRating: MealRating = {
+                mealId: mockMealId,
+                rating: 8.14,
+                count: 123
+            };
+
+            jest.spyOn(redisService, 'hasMeal').mockResolvedValueOnce(mockHasMeal);
+            jest.spyOn(mealRatingRepository, 'getAverageRatingForMeal').mockResolvedValueOnce(mockMealRating);
+
+            return request(app.getHttpServer())
+                .get(`/meals/${mockMealId}/rating`)
+                .expect(200)
+                .expect(mockMealRating);
+        });
+    });
+
     describe('/meals/:id/comment (POST)', () => {
         it('should post a new comment for a particular meal', () => {
             const mockMealId = 'mock meal id';
@@ -250,6 +278,39 @@ describe('UserController (e2e)', () => {
                 .set('Authorization', 'Bearer token')
                 .send(mockRequestBody)
                 .expect(201);
+        });
+    });
+
+    describe('/meals/:id/rating (POST)', () => {
+        it('should post a new rating for a specific meal', () => {
+            const mockMealId = 'mock meal id';
+            const mockRequestBody = {
+                mealId: mockMealId,
+                user: 'mock user name',
+                rating: 10
+            } as any;
+            const mockMealRating = {
+                ...mockRequestBody,
+                posted: Date.now()
+            } as any;
+            const mockUser = {
+                _id: '635981f6e40f61599e839ddb',
+                login: 'user',
+                password: 'hashed'
+            } as any;
+
+            jest.spyOn(jwtManagerService, 'verifyAccessToken').mockResolvedValue(mockUser);
+            jest.spyOn(redisService, 'getAccessToken').mockResolvedValue('token');
+            jest.spyOn(redisService, 'hasMeal').mockResolvedValueOnce(true);
+            jest.spyOn(mealRatingRepository, 'create').mockResolvedValueOnce(mockMealRating);
+
+            return request(app.getHttpServer())
+                .post(`/meals/${mockMealId}/rating`)
+                .set('Cookie', ['accessToken=token'])
+                .set('Accept', 'application/json')
+                .set('Authorization', 'Bearer token')
+                .send(mockRequestBody)
+                .expect(200);
         });
     });